pyproject.toml
[project]
name = "playground-agent"
version = "0.1.0"
description = "Extensible insurance consultation agent (LangGraph)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langchain-core>=0.3,<0.4",
  "langchain-openai>=0.2,<0.4",
  "langgraph>=0.2,<0.4",
  "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = ["ruff>=0.6"]

[tool.ruff]
line-length = 100
target-version = "py310"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["insurance_agent*"]
